requirements.txt
selenium==3.141.0
slackclient==2.9.3